relocated index page icons folder in Index folder
