+++ src/components/index/icons/problems-to-solve-icons/CommissionsIcon.tsx
