+++ src/components/index/icons/problems-to-solve-icons/TransparencyIcon.tsx
