+++ src/components/index/icons/support-us-icons/PartnershipIcon.tsx
